+++ src/kit/src/main/java/io/ayte/utility/supplier/kit/factory/LinkedListFactory.java
@@ -1,16 +1,16 @@
 package io.ayte.utility.supplier.kit.factory;
 
+import io.ayte.utility.supplier.ThreadSafeSupplier;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import java.util.LinkedList;
 import java.util.List;
-import java.util.function.Supplier;
 
 @ToString
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class LinkedListFactory<T> implements Supplier<List<T>> {
+public class LinkedListFactory<T> implements ThreadSafeSupplier<List<T>> {
     private static final LinkedListFactory INSTANCE = new LinkedListFactory<>();
 
     @Override