+++ src/kit/src/main/java/io/ayte/utility/supplier/kit/factory/TreeSetFactory.java
@@ -1,15 +1,15 @@
 package io.ayte.utility.supplier.kit.factory;
 
+import io.ayte.utility.supplier.ThreadSafeSupplier;
 import lombok.NonNull;
 import lombok.ToString;
 
 import java.util.Comparator;
 import java.util.NavigableSet;
 import java.util.TreeSet;
-import java.util.function.Supplier;
 
 @ToString
-public class TreeSetFactory<T> implements Supplier<NavigableSet<T>> {
+public class TreeSetFactory<T> implements ThreadSafeSupplier<NavigableSet<T>> {
     private final Comparator<? super T> comparator;
 
     private TreeSetFactory(Comparator<? super T> comparator) {