+++ src/kit/src/main/java/io/ayte/utility/supplier/kit/standard/ConstantSupplier.java
@@ -1,16 +1,16 @@
-package io.ayte.utility.supplier.kit.simple;
+package io.ayte.utility.supplier.kit.standard;
 
+import io.ayte.utility.supplier.StableSupplier;
+import io.ayte.utility.supplier.ThreadSafeSupplier;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
-import java.util.function.Supplier;
-
 @ToString
 @EqualsAndHashCode
 @RequiredArgsConstructor(access = AccessLevel.PRIVATE)
-public class ConstantSupplier<T> implements Supplier<T> {
+public class ConstantSupplier<T> implements StableSupplier<T>, ThreadSafeSupplier<T> {
     private final T value;
 
     @Override